Feat: User모듈과 Auth 파일 재정비
