+++ src/api/auth/auth.module.ts
@@ -2,16 +2,23 @@ import { Module } from '@nestjs/common';
 import { PassportModule } from '@nestjs/passport';
 import { JwtModule } from '@nestjs/jwt';
 import { AuthService } from './auth.service';
+import { UserModule } from '../user/user.module';
+import { JWTStrategy } from './strategy/jwt.strategy';
+import { UserService } from '../user/user.service';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { UserEntity } from '../user/entities/user.entity';
 
 @Module({
   imports: [
-    PassportModule.register({ defaultStrategy: 'jwt' }),
+    TypeOrmModule.forFeature([UserEntity]),
+    UserModule,
+    PassportModule,
     JwtModule.register({
       secret: process.env.JWT_SECRET_KEY,
-      signOptions: { expiresIn: process.env.JWT_SECRET_KEY },
+      signOptions: { expiresIn: parseInt(process.env.JWT_SECRET_KEY) },
     }),
   ],
-  providers: [AuthService],
+  providers: [AuthService, JWTStrategy, UserService],
   exports: [AuthService],
 })
 export class AuthModule {}
