+++ src/api/user/user.controller.ts
@@ -7,17 +7,21 @@ import {
   ApiTags,
   ApiUnauthorizedResponse,
 } from '@nestjs/swagger';
+import { AuthService } from '../auth/auth.service';
 import { User } from './decorator/user.decorator';
 import { UserAPIDocs } from './docs/user.docs';
 import { CreateUserDto } from './dtos/createUser.dto';
 import { LoginRequestDto } from './dtos/loginRequest.dto';
-import { LocalAuthGuard } from './guard/local.guard';
+import { LocalAuthGuard } from '../auth/guard/local.guard';
 import { UserService } from './user.service';
 
 @ApiTags('User API')
 @Controller('user')
 export class UserController {
-  constructor(private readonly userService: UserService) {}
+  constructor(
+    private readonly userService: UserService,
+    private readonly authService: AuthService,
+  ) {}
 
   // 회원가입
   @ApiOperation(UserAPIDocs.signUpOperation())
@@ -32,9 +36,10 @@ export class UserController {
   @ApiOperation(UserAPIDocs.loginOperation())
   @ApiResponse(UserAPIDocs.loginResponse())
   @ApiUnauthorizedResponse(UserAPIDocs.loginUnauthorizedResponse())
+  @UseGuards(LocalAuthGuard)
   @Post('/login')
   login(@Body() loginRequestDto: LoginRequestDto) {
-    return this.userService.login(loginRequestDto);
+    return this.authService.login(loginRequestDto);
   }
 
   // 회원탈퇴
