+++ src/api/user/user.service.ts
@@ -10,14 +10,14 @@ import { UserEntity } from './entities/user.entity';
 import * as bcrypt from 'bcryptjs';
 import { Status } from './entities/status.enum';
 import { LoginRequestDto } from './dtos/loginRequest.dto';
-import { AuthService } from '../auth/auth.service';
+// import { AuthService } from '../auth/auth.service';
 
 @Injectable()
 export class UserService {
   constructor(
     @InjectRepository(UserEntity)
     private userRepository: Repository<UserEntity>,
-    private readonly authService: AuthService,
+    // private readonly authService: AuthService,
   ) {}
 
   async signup(createUserDto: CreateUserDto) {
@@ -43,29 +43,17 @@ export class UserService {
     return result;
   }
 
-  async validateUser(email: string, password: string): Promise<any> {
-    const user = await this.userRepository.findOne({ where: { email } });
-
-    if (!user) throw new UnauthorizedException('존재하지 않는 계정입니다.');
-
-    if (user && (await bcrypt.compare(password, user.password))) {
-      delete user.password;
-      const token = await this.authService.generateAccessToken(user);
-      return { user, token };
-    } else throw new UnauthorizedException('비밀번호가 틀립니다.');
-  }
-
-  async login(loginRequestDto: LoginRequestDto) {
-    const { email, password } = loginRequestDto;
-    const result = await this.validateUser(email, password);
-    return result;
-  }
-
   async deleteUserByEmail(id: number) {
     const found = await this.userRepository.findOne({ where: { id } });
     found.deleteAt = new Date();
     found.status = Status.STOP;
     await this.userRepository.save(found);
     return { success: true };
+  }
+
+  async findByEmail(email: string) {
+    const user = await this.userRepository.findOne({ where: { email } });
+    if (!user) throw new UnauthorizedException('존재하지 않는 계정입니다.');
+    return user;
   }
 }
